Split off shader cache from pipeline cache
