+++ crates/bevy_render/src/render_resource/mod.rs
@@ -12,6 +12,7 @@ mod pipeline_cache;
 mod pipeline_specializer;
 pub mod resource_macros;
 mod shader;
+mod shader_cache;
 mod specializer;
 mod storage_buffer;
 mod texture;
@@ -29,6 +30,7 @@ pub use pipeline::*;
 pub use pipeline_cache::*;
 pub use pipeline_specializer::*;
 pub use shader::*;
+pub use shader_cache::*;
 pub use specializer::*;
 pub use storage_buffer::*;
 pub use texture::*;
